Fix Windows tests after removing winpty
With winpty, `build/Release/pty.node` is created, which happens to be
the same file name as on Linux and macOS, which resulted in that the
tests happened to work. Since removing winpty, that file is not created
anymore, so we need to fix the tests to not run unixTerminal on Windows.